requirements.txt
pipreqs==0.4.11
pytest==7.1.2
pytest-html==3.1.1
pytest-xdist==2.5.0
selenium==4.3.0
webdriver-manager==3.8.1
PyYAML~=6.0